FIX Errors: WPLAB6

+++ wp-starter-core/src/main/java/mk/ukim/finki/wp/service/IGroupService.java
@@ -9,8 +9,8 @@ import java.util.List;
  */
 public interface IGroupService {
     List<Group> findAll();
-    Group findById(Long id);
+    Group findById(Integer id);
     void save(Group entity);
-    void update(Long id, Group entity);
-    void delete(Long id);
+    void update(Integer id, Group entity);
+    void delete(Integer id);
 }
